chore: split internal project view into new components
