+++ frontend/src/views/internal/ProjectView.vue
@@ -1,35 +1,17 @@
 <script setup lang="ts">
 import { storeToRefs } from 'pinia';
-import { filesize } from 'filesize';
 import { useI18n } from 'vue-i18n';
 import { computed, onBeforeMount, provide, ref } from 'vue';
 import { useRouter } from 'vue-router';
 
-import AuthorizationCard from '@/components/authorization/AuthorizationCard.vue';
-import AuthorizationCardLite from '@/components/authorization/AuthorizationCardLite.vue';
-import DeleteDocument from '@/components/file/DeleteDocument.vue';
-import DocumentCard from '@/components/file/DocumentCard.vue';
-import ElectrificationProjectForm from '@/components/electrification/project/ProjectFormNavigator.vue';
-import EnquiryCard from '@/components/enquiry/EnquiryCard.vue';
-import FileUpload from '@/components/file/FileUpload.vue';
-import { default as GeneralProjectForm } from '@/components/general/project/ProjectFormNavigator.vue';
-import { default as HousingProjectForm } from '@/components/housing/project/ProjectFormNavigator.vue';
-import NoteHistoryCard from '@/components/note/NoteHistoryCard.vue';
+import ProjectAuthorizationsTab from '@/components/projectCommon/ProjectAuthorizationsTab.vue';
+import ProjectEnquiryTab from '@/components/projectCommon/ProjectEnquiryTab.vue';
+import ProjectFilesTab from '@/components/projectCommon/ProjectFilesTab.vue';
+import ProjectInformationTab from '@/components/projectCommon/ProjectInformationTab.vue';
+import ProjectNotesTab from '@/components/projectCommon/ProjectNotesTab.vue';
+import ProjectRoadmapTab from '@/components/projectCommon/ProjectRoadmapTab.vue';
 import ProjectTeamTab from '@/components/projectCommon/ProjectTeamTab.vue';
-import Roadmap from '@/components/roadmap/Roadmap.vue';
-import {
-  Button,
-  Column,
-  DataTable,
-  IconField,
-  InputIcon,
-  InputText,
-  Tab,
-  Tabs,
-  TabList,
-  TabPanel,
-  TabPanels
-} from '@/lib/primevue';
+import { Button, Tab, Tabs, TabList, TabPanel, TabPanels } from '@/lib/primevue';
 import {
   activityContactService,
   documentService,
@@ -44,14 +26,12 @@ import {
 } from '@/services';
 import { useAppStore, useAuthZStore, usePermitStore, useProjectStore } from '@/store';
 import { Action, Initiative, Resource, RouteName } from '@/utils/enums/application';
-import { ApplicationStatus } from '@/utils/enums/projectCommon';
-import { formatDateLong } from '@/utils/formatters';
-import { projectServiceKey } from '@/utils/keys';
+import { projectAuthorizationRouteNameKey, projectNoteRouteNameKey, projectServiceKey } from '@/utils/keys';
 import { generalErrorHandler, getFilenameAndExtension } from '@/utils/utils';
 
 import type { Ref } from 'vue';
 import type { IDraftableProjectService } from '@/interfaces/IProjectService';
-import type { Document, ElectrificationProject, Enquiry, GeneralProject, HousingProject, User } from '@/types';
+import type { Document, Enquiry, User } from '@/types';
 
 // Props
 const { initialTab = '0', projectId } = defineProps<{
@@ -93,18 +73,6 @@ const HOUSING_INITIATIVE_STATE: InitiativeState = {
   projectService: housingProjectService
 };
 
-const SORT_ORDER = {
-  ASCENDING: 1,
-  DESCENDING: -1
-};
-const SORT_TYPES = {
-  CREATED_AT: 'createdAt',
-  FILENAME: 'filename',
-  FILESIZE: 'filesize',
-  MIME_TYPE: 'mimeType',
-  CREATED_BY: 'createdByFullName'
-};
-
 // Composables
 const { t } = useI18n();
 const router = useRouter();
@@ -116,13 +84,9 @@ const projectStore = useProjectStore();
 const { getPermitTypes } = storeToRefs(permitStore);
 const {
   getActivityContacts,
-  getAuthsCompleted,
-  getAuthsNeeded,
-  getAuthsNotNeeded,
-  getAuthsOnGoing,
-  getAuthsUnderInvestigation,
   getDocuments,
   getProject,
+  getProjectIsCompleted,
   getNoteHistory,
   getPermits,
   getRelatedEnquiries
@@ -135,80 +99,16 @@ const initiativeState: Ref<InitiativeState> = ref(HOUSING_INITIATIVE_STATE);
 const liveName: Ref<string> = ref('');
 const loading: Ref<boolean> = ref(true);
 const noteHistoryCreatedByFullnames: Ref<{ noteHistoryId: string; createdByFullname: string }[]> = ref([]);
-const gridView: Ref<boolean> = ref(false);
-const searchTag: Ref<string> = ref('');
-const sortOrder: Ref<number | undefined> = ref(Number(SORT_ORDER.DESCENDING));
-const sortType: Ref<string> = ref(SORT_TYPES.CREATED_AT);
-
-const filteredDocuments = computed(() => {
-  let tempDocuments = getDocuments.value;
-  tempDocuments = tempDocuments.filter((x) => {
-    return searchTag.value ? x.filename.toLowerCase().includes(searchTag.value.toLowerCase()) : x;
-  });
-  switch (sortType.value) {
-    case SORT_TYPES.FILENAME:
-      tempDocuments = tempDocuments.sort((a, b) =>
-        sortComparator(sortOrder.value, a.filename.toLowerCase(), b.filename.toLowerCase())
-      );
-      break;
-    case SORT_TYPES.CREATED_AT:
-      tempDocuments = tempDocuments.sort((a, b) => sortComparator(sortOrder.value, a.createdAt ?? 0, b.createdAt ?? 0));
-      break;
-    case SORT_TYPES.FILESIZE:
-      tempDocuments = tempDocuments.sort((a, b) => sortComparator(sortOrder.value, a.filesize, b.filesize));
-      break;
-    case SORT_TYPES.MIME_TYPE:
-      tempDocuments = tempDocuments.sort((a, b) =>
-        sortComparator(sortOrder.value, a.mimeType.toLowerCase(), b.mimeType.toLowerCase())
-      );
-      break;
-    case SORT_TYPES.CREATED_BY:
-      tempDocuments = tempDocuments.sort((a, b) =>
-        sortComparator(sortOrder.value, a.createdByFullName.toLowerCase(), b.createdByFullName.toLowerCase())
-      );
-      break;
-  }
-  return tempDocuments;
-});
-
-const isCompleted = computed(() => {
-  return getProject.value?.applicationStatus === ApplicationStatus.COMPLETED;
-});
 
 // Providers
+const provideProjectNoteRouteName = computed(() => initiativeState.value.projectNoteRouteName);
+const provideProjectAuthorizationRouteName = computed(() => initiativeState.value.projectAuthorizationRouteName);
 const provideProjectService = computed(() => initiativeState.value.projectService);
+provide(projectNoteRouteNameKey, provideProjectNoteRouteName);
+provide(projectAuthorizationRouteNameKey, provideProjectAuthorizationRouteName);
 provide(projectServiceKey, provideProjectService);
 
 // Actions
-function sortComparator(sortValue: number | undefined, a: string | number, b: string | number) {
-  if (sortValue === SORT_ORDER.ASCENDING) return a > b ? 1 : -1;
-  else return a < b ? 1 : -1;
-}
-
-function toAuthorization(authId: string) {
-  router.push({
-    name: initiativeState.value.projectAuthorizationRouteName,
-    params: {
-      permitId: authId,
-      projectId: projectId
-    }
-  });
-}
-
-function toEditNote(noteHistoryId: string) {
-  router.push({
-    name: initiativeState.value.projectNoteRouteName,
-    params: {
-      noteHistoryId: noteHistoryId,
-      projectId: projectId
-    }
-  });
-}
-
-function updateLiveName(name: string) {
-  liveName.value = name;
-}
-
 onBeforeMount(async () => {
   try {
     switch (getInitiative.value) {
@@ -300,7 +200,7 @@ onBeforeMount(async () => {
         {{ getProject.activityId }}
       </span>
       <span
-        v-if="isCompleted"
+        v-if="getProjectIsCompleted"
         class="ml-0"
       >
         (Completed)
@@ -364,407 +264,28 @@ onBeforeMount(async () => {
     </TabList>
     <TabPanels>
       <TabPanel :value="0">
-        <span v-if="!loading && getProject">
-          <HousingProjectForm
-            v-if="getInitiative === Initiative.HOUSING"
-            :editable="!isCompleted && useAuthZStore().can(getInitiative, Resource.HOUSING_PROJECT, Action.UPDATE)"
-            :project="getProject as HousingProject"
-            @input-project-name="updateLiveName"
-          />
-          <GeneralProjectForm
-            v-if="getInitiative === Initiative.GENERAL"
-            :editable="!isCompleted && useAuthZStore().can(getInitiative, Resource.GENERAL_PROJECT, Action.UPDATE)"
-            :project="getProject as GeneralProject"
-            @input-project-name="updateLiveName"
-          />
-          <ElectrificationProjectForm
-            v-if="getInitiative === Initiative.ELECTRIFICATION"
-            :editable="
-              !isCompleted && useAuthZStore().can(getInitiative, Resource.ELECTRIFICATION_PROJECT, Action.UPDATE)
-            "
-            :project="getProject as ElectrificationProject"
-            @input-project-name="updateLiveName"
-          />
-        </span>
+        <ProjectInformationTab v-model:live-name="liveName" />
       </TabPanel>
       <TabPanel :value="1">
-        <div class="mb-4 border-dashed file-upload rounded-md">
-          <FileUpload
-            v-if="activityId"
-            :activity-id="activityId"
-            :disabled="isCompleted || !useAuthZStore().can(getInitiative, Resource.DOCUMENT, Action.CREATE)"
-          />
-        </div>
-        <div class="flex flex-row justify-between pb-4">
-          <div class="flex items-center">
-            <IconField icon-position="left">
-              <InputIcon class="pi pi-search" />
-              <InputText
-                v-model="searchTag"
-                placeholder="Search"
-              />
-            </IconField>
-          </div>
-          <div class="items-end">
-            <Button
-              aria-label="List"
-              class="view-switch-button"
-              @click="gridView = false"
-            >
-              <font-awesome-icon
-                icon="fa-solid fa-list"
-                class="fa-lg"
-                :class="gridView ? 'list-grid-deselected-icon' : 'list-grid-selected-icon'"
-              />
-            </Button>
-            <Button
-              aria-label="Grid"
-              class="view-switch-button"
-              @click="gridView = true"
-            >
-              <font-awesome-icon
-                icon="fa-solid fa-grip"
-                class="fa-lg"
-                :class="gridView ? 'list-grid-selected-icon' : 'list-grid-deselected-icon'"
-              />
-            </Button>
-          </div>
-        </div>
-        <div
-          v-if="gridView"
-          class="grid grid-cols-12 gap-4 nested-grid"
-        >
-          <DataTable
-            v-if="gridView"
-            class="remove-padding col-span-12"
-            removable-sort
-            :sort-field="SORT_TYPES.CREATED_AT"
-            :sort-order="SORT_ORDER.DESCENDING"
-            @update:sort-order="
-              (order: number | undefined) => {
-                sortOrder = order ?? SORT_ORDER.DESCENDING;
-              }
-            "
-            @update:sort-field="
-              (field: string) => {
-                sortType = field;
-              }
-            "
-          >
-            <template #empty>
-              <div class="flex justify-center" />
-            </template>
-            <Column
-              sortable
-              field="filename"
-              header="File name"
-              class="w-36"
-            />
-            <Column
-              field="createdAt"
-              sortable
-              header="Upload date"
-              class="w-40"
-            />
-            <Column
-              field="filesize"
-              sortable
-              header="Size"
-              class="w-24"
-            />
-            <Column
-              field="mimeType"
-              sortable
-              header="Type"
-              class="w-40"
-            />
-            <Column />
-          </DataTable>
-
-          <div class="col-span-12">
-            <div class="grid grid-cols-6 gap-4">
-              <div
-                v-for="(document, index) in filteredDocuments"
-                :key="document.documentId"
-                :index="index"
-                class="col-span-1"
-              >
-                <DocumentCard
-                  :document="document"
-                  :editable="!isCompleted"
-                  class="hover-hand hover-shadow"
-                  @click="documentService.downloadDocument(document.documentId, document.filename)"
-                />
-              </div>
-            </div>
-          </div>
-        </div>
-        <DataTable
-          v-if="!gridView"
-          :value="filteredDocuments"
-          removable-sort
-          :sort-field="SORT_TYPES.CREATED_AT"
-          :sort-order="SORT_ORDER.DESCENDING"
-          :row-hover="true"
-        >
-          <Column
-            field="filename"
-            header="File name"
-            sortable
-          >
-            <template #body="{ data }">
-              <a
-                href="#"
-                @click="
-                  () => {
-                    if (useAuthZStore().can(getInitiative, Resource.DOCUMENT, Action.READ))
-                      documentService.downloadDocument(data.documentId, data.filename);
-                  }
-                "
-              >
-                {{ data.filename }}
-              </a>
-            </template>
-          </Column>
-          <Column
-            field="createdAt"
-            header="Upload date"
-            sortable
-          >
-            <template #body="{ data }">
-              {{ formatDateLong(data.createdAt) }}
-            </template>
-          </Column>
-          <Column
-            field="filesize"
-            header="Size"
-            sortable
-          >
-            <template #body="{ data }">
-              {{ filesize(data.filesize) }}
-            </template>
-          </Column>
-          <Column
-            field="extension"
-            header="Type"
-            sortable
-          />
-          <Column
-            field="createdByFullName"
-            header="Uploaded by"
-            sortable
-          />
-          <Column field="fileAction">
-            <template #header>
-              <div class="flex justify-center w-full">
-                <b>Action</b>
-              </div>
-            </template>
-            <template #body="{ data }">
-              <div class="flex justify-center">
-                <DeleteDocument
-                  :disabled="isCompleted || !useAuthZStore().can(getInitiative, Resource.DOCUMENT, Action.DELETE)"
-                  :document="data"
-                />
-              </div>
-            </template>
-          </Column>
-        </DataTable>
+        <ProjectFilesTab />
       </TabPanel>
       <TabPanel :value="2">
-        <div class="flex items-center pb-5">
-          <div class="grow">
-            <p class="font-bold">{{ t('views.i.projectView.applicableAuthorizations') }} ({{ getPermits.length }})</p>
-          </div>
-          <Button
-            data-test-id="add-authorization-button"
-            :aria-label="t('views.i.projectView.addAuthorization')"
-            :disabled="isCompleted || !useAuthZStore().can(getInitiative, Resource.PERMIT, Action.CREATE)"
-            @click="
-              router.push({
-                name: initiativeState.projectAddAuthorizationRouteName,
-                params: {
-                  projectId: projectId
-                }
-              })
-            "
-          >
-            <font-awesome-icon
-              class="pr-2"
-              icon="fa-solid fa-plus"
-            />
-            {{ t('views.i.projectView.addAuthorization') }}
-          </Button>
-        </div>
-        <!-- On going Authorizations -->
-        <div
-          v-for="(permit, index) in getAuthsOnGoing"
-          :id="permit.permitId"
-          :key="permit.permitId"
-          :index="index"
-          class="mb-6 mt-6"
-        >
-          <AuthorizationCard
-            :editable="!isCompleted"
-            :permit="permit"
-            @authorization-card:more="toAuthorization(permit.permitId)"
-          />
-        </div>
-        <!-- Authorizations with needed = under investigation -->
-        <div
-          v-if="getAuthsUnderInvestigation.length > 0"
-          class="mb-8 mt-16"
-        >
-          <h4 class="mb-6">{{ t('views.i.projectView.underInvestigation') }}</h4>
-          <div
-            v-for="(permit, index) in getAuthsUnderInvestigation"
-            :id="permit.permitId"
-            :key="permit.permitId"
-            :index="index"
-            class="my-2"
-          >
-            <AuthorizationCardLite
-              :editable="!isCompleted"
-              :permit="permit"
-              @authorization-card-lite:more="toAuthorization(permit.permitId)"
-            />
-          </div>
-        </div>
-        <!-- Authorizations with needed = Yes & stage = Pre-submission -->
-        <div
-          v-if="getAuthsNeeded.length > 0"
-          class="mb-8 mt-16"
-        >
-          <h4 class="mb-6">{{ t('views.i.projectView.needed') }}</h4>
-          <div
-            v-for="(permit, index) in getAuthsNeeded"
-            :id="permit.permitId"
-            :key="permit.permitId"
-            :index="index"
-            class="my-2"
-          >
-            <AuthorizationCardLite
-              :editable="!isCompleted"
-              :permit="permit"
-              @authorization-card-lite:more="toAuthorization(permit.permitId)"
-            />
-          </div>
-        </div>
-        <!--Authorizations when its state=Approved, Denied, Cancelled, OR Withdrawn.-->
-        <div
-          v-if="getAuthsCompleted.length > 0"
-          class="mb-8 mt-16"
-        >
-          <h4 class="mb-6">{{ t('views.i.projectView.completed') }}</h4>
-          <div
-            v-for="(permit, index) in getAuthsCompleted"
-            :id="permit.permitId"
-            :key="permit.permitId"
-            :index="index"
-            class="my-2"
-          >
-            <AuthorizationCard
-              :editable="!isCompleted"
-              :permit="permit"
-              @authorization-card:more="toAuthorization(permit.permitId)"
-            />
-          </div>
-        </div>
-        <!--Authorizations when needed = NO-->
-        <div
-          v-if="getAuthsNotNeeded.length > 0"
-          class="mb-8 mt-16"
-        >
-          <h4 class="mb-6">{{ t('views.i.projectView.notNeeded') }}</h4>
-          <div
-            v-for="(permit, index) in getAuthsNotNeeded"
-            :id="permit.permitId"
-            :key="permit.permitId"
-            :index="index"
-            class="my-2"
-          >
-            <AuthorizationCardLite
-              :editable="!isCompleted"
-              :permit="permit"
-              @authorization-card-lite:more="toAuthorization(permit.permitId)"
-            />
-          </div>
-        </div>
+        <ProjectAuthorizationsTab />
       </TabPanel>
       <TabPanel :value="3">
-        <div class="flex items-center pb-5">
-          <div class="grow">
-            <p class="font-bold">Notes ({{ getNoteHistory.length }})</p>
-          </div>
-          <Button
-            aria-label="Add note"
-            :disabled="isCompleted || !useAuthZStore().can(getInitiative, Resource.NOTE, Action.CREATE)"
-            @click="
-              router.push({
-                name: initiativeState.projectNoteRouteName,
-                params: {
-                  projectId: projectId
-                }
-              })
-            "
-          >
-            <font-awesome-icon
-              class="pr-2"
-              icon="fa-solid fa-plus"
-            />
-            Add note
-          </Button>
-        </div>
-        <div v-if="!loading">
-          <div
-            v-for="(noteHistory, index) in getNoteHistory"
-            :key="noteHistory.noteHistoryId"
-            :index="index"
-            class="mb-6"
-          >
-            <NoteHistoryCard
-              :editable="!isCompleted"
-              :note-history="noteHistory"
-              :created-by-full-name="
-                noteHistoryCreatedByFullnames.find((x) => x.noteHistoryId === noteHistory.noteHistoryId)
-                  ?.createdByFullname
-              "
-              @edit-note-history="(e) => toEditNote(e)"
-              @delete-note-history="(e) => projectStore.removeNoteHistory(e)"
-              @update-note-history="(e) => projectStore.updateNoteHistory(e)"
-            />
-          </div>
-        </div>
+        <ProjectNotesTab />
       </TabPanel>
       <TabPanel :value="4">
-        <Roadmap
-          v-if="!loading && activityId"
-          :activity-id="activityId"
-          :editable="!isCompleted && useAuthZStore().can(getInitiative, Resource.ROADMAP, Action.CREATE)"
+        <ProjectRoadmapTab
+          v-if="getProject"
+          :editable="!getProjectIsCompleted && useAuthZStore().can(getInitiative, Resource.ROADMAP, Action.CREATE)"
         />
       </TabPanel>
       <TabPanel :value="5">
-        <div class="flex items-center pb-2">
-          <div class="grow">
-            <p class="font-bold">
-              {{ t('views.i.projectView.tabRelatedEnquiries') }} ({{ getRelatedEnquiries.length }})
-            </p>
-          </div>
-        </div>
-        <div
-          v-for="(enquiry, index) in getRelatedEnquiries"
-          :key="enquiry.enquiryId"
-          :index="index"
-          class="col-span-12 mb-6"
-        >
-          <EnquiryCard :enquiry="enquiry" />
-        </div>
+        <ProjectEnquiryTab />
       </TabPanel>
       <TabPanel :value="6">
-        <ProjectTeamTab
-          v-if="getProject"
-          :activity-id="getProject.activityId"
-        />
+        <ProjectTeamTab v-if="getProject" />
       </TabPanel>
     </TabPanels>
   </Tabs>
@@ -778,41 +299,9 @@ onBeforeMount(async () => {
   }
 }
 
-:deep(.remove-padding.p-datatable .p-datatable-tbody > tr > td) {
-  display: none;
-}
-
 :deep(.p-tab) {
   &.no-underline {
     text-decoration: none;
-  }
-}
-
-.p-button.p-component.view-switch-button {
-  background-color: transparent;
-  border: none;
-  padding-left: 0;
-}
-
-.list-grid-selected-icon {
-  color: var(--p-primary-color);
-}
-
-.list-grid-deselected-icon {
-  color: var(--p-content-hover-background);
-  &:hover {
-    color: var(--primary-hover-color);
-  }
-}
-
-.p-inputtext.p-component {
-  width: 20rem;
-}
-
-.file-upload {
-  color: var(--p-greyscale-500);
-  &:hover {
-    color: var(--p-content-hover-background);
   }
 }
 </style>
